DEVELOP: `Layout.jsx` component
A layout component that will be re-used by container components.

According to a StackOverflow post regarding Next.js and global components, it is better to explicitly wrap the pages/containers that will use the layout instead of wrapping the entire app due to code-splitting issues. I'm not sure if this is legit or not but it makes sense from an SSR standpoint to me.

Link to StackOverflow post: https://stackoverflow.com/a/60778096